Created SurveyCode component
