+++ frontend/src/App.jsx
@@ -34,4 +34,4 @@ const App = () => {
     )
 }
 
-export default App
+export default App;
