+++ frontend/src/welcome/Welcome.jsx
@@ -1,35 +1,6 @@
-import {useContext, useState} from 'react';
-import {Button} from 'primereact/button';
-import {InputText} from 'primereact/inputtext';
-import {getSurvey} from "../api/surveyApi.js";
-import {SurveyContext} from "../context/SurveyContext.jsx";
+import SurveyCode from "../home/SurveyCode.jsx";
 
 const Welcome = ({toast}) => {
-    const [survey, setSurvey] = useContext(SurveyContext);
-    const [code, setCode] = useState((window.location.pathname).toString().substring(1));
-
-    const codeSubmit = async () => {
-        if (code) {
-            const surveyStatus = await getSurvey(code);
-            if (surveyStatus.success) {
-                setSurvey(surveyStatus.body);
-            } else {
-                toast.current.show({
-                    severity: 'error',
-                    life: 3000,
-                    summary: 'Survey Error',
-                    detail: surveyStatus.body.message,
-                });
-            }
-        } else {
-            toast.current.show({
-                severity: 'error',
-                life: 3000,
-                summary: 'Survey Error',
-                detail: 'Survey code is required',
-            });
-        }
-    }
 
     return (
         <>
@@ -42,10 +13,7 @@ const Welcome = ({toast}) => {
                     <p className="text-2xl">Polls made simple.</p>
                 </div>
                 <div className="flex gap-2">
-                    <InputText value={code} onChange={(e) => setCode(e.target.value)} placeholder="Enter Code"
-                               className="p-inputtext-lg"/>
-                    <Button icon="pi pi-arrow-right" size="large" style={{boxShadow: "none"}}
-                            onClick={() => codeSubmit()}/>
+                    <SurveyCode toast={toast}/>
                 </div>
             </div>
         </>
